requirements.txt
nicegui
requests
cairosvg
python-dotenv
supabase
weasyprint
email-validator